WIP: Transfer of necessary functionality from Components to GameObject
